refactor: useState로 관리하던 params 전부 개선하여 적용
